requirements.txt
Django==3.2
dj-database-url==0.5.0
dj-static==0.0.6
gunicorn==19.8.1
psycopg2==2.8.6
static==1.1.1
requests
djangorestframework==3.14.0
django-allauth==0.51.0
django-storages==1.11.1
boto3==1.17.58
Pillow==9.3.0
sendgrid-django==4.2.0
whitenoise==5.0.1
git+https://github.com/JMIdeaMaker/django-vote.git@master
django-recaptcha==3.0.0
git+https://github.com/jmideamaker/django-messages.git@master
martor==1.6.15
django-s3direct
allcities==1.0.3